Website for Agones
This is a new static site for Agones, as well as the tools to enable people to
write documentation for new features, without it displaying until the feature
is released.

Integration with cloudbuild for previews is also build in (preview is the
current documentation + all the documentation for the new features), as well
as a link checker that runs as part of the test suite.

This also brings into the codebase all the redirect code for go-get as
well as the chart redirects as well.

Closes #410